Update reftest to always generate a directory at the root of the reference model (#1219)
This change impacts only Mountpoint's reference tests.

The current strategy for generating the tree allows for the root to be a
file, which is not possible. This leads to us adding special cases to
the reftest comparison logic as well as having bizarre test cases which
are hard to understand.

This change updates the strategy by ensuring that the root is always a
directory, and simplifies some of the unused proptest layers.

### Does this change impact existing behavior?

No change to existing Mountpoint behavior. This changes the type of
trees generated by our reference tests, removing those that are not
possible in Mountpoint.

### Does this change need a changelog entry?

No.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Signed-off-by: Daniel Carl Jones <djonesoa@amazon.com>