Fix Clippy warnings on Rust 1.84 (#1216)
Rust 1.84 has just released and with it, new clippy lints have been
added.

We're also updating to Rust 1.84 explicitly rather than returning to
stable to avoid the issues we had with the most recent release

### Does this change impact existing behavior?

No

### Does this change need a changelog entry?

No

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Simon Beal <simobeal@amazon.com>