Increment version numbers for next release and update guidance (#1229)
Set up new guidance for incrementing version numbers and implement it
for Mountpoint and the client crates.

Under the new guidance, the patch version will be incremented
immediately after releasing Mountpoint or publishing the crates, so that
development on `main` continues under a new provisional version.
When new features / breaking changes are introduced, the version number
will be contextually incremented as the changes are documented in the
changelog.

### Does this change impact existing behavior?

No

### Does this change need a changelog entry?

No

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Alessandro Passaro <alexpax@amazon.co.uk>